Next update to 12.2.3 (#731)
* fix build

* include next base config

* add commit

* fix build
