+++ files_to_zip.txt
@@ -4,9 +4,10 @@ Dockerrun.aws.json
 .env
 .env.staging
 .next
+next.base.js
 next.config.js
 node_modules
 package.json
 public
 prisma
-seedData/space/space-da74cab3-c2b6-40bb-8734-0de5375b0fce-pages-1657887621286
+seedData
