+++ next.base.js
@@ -1,0 +1,26 @@
+
+const esmModules = [
+  '@bangle.dev/base-components',
+  '@bangle.dev/core',
+  '@bangle.dev/pm',
+  '@bangle.dev/react',
+  '@bangle.dev/utils',
+  '@bangle.dev/markdown',
+  '@bangle.dev/tooltip',
+  '@bangle.dev/react-menu',
+  '@bangle.dev/table',
+  '@popperjs/core',
+  '@fullcalendar/common',
+  '@fullcalendar/core',
+  '@fullcalendar/daygrid',
+  '@fullcalendar/interaction',
+  '@fullcalendar/react',
+  'react-dnd',
+  // 'react-dnd-scrolling',
+  'react-pdf',
+  '@hookform/resolvers',
+  'lit-share-modal-v3-react-17',
+  'uuid'
+];
+
+exports.esmModules = esmModules;
